* Fixed up a bunch more warnings + Reimplemented copy/paste of map architecture
git-svn-id: http://mancubus.net/svn/hosted/slade/slade_20_alpha@78 1ddb354d-3e0c-0410-833a-d70222086f97

+++ src/bsp.cpp
@@ -459,5 +459,5 @@ void update_visibility()
 	memset(ssects_visible, 0, gl_ssects.size());
 	set_visbuffer(0);
 	block_behind();
-	walk_bsp_tree(gl_nodes.size() - 1);
+	walk_bsp_tree((DWORD)gl_nodes.size() - 1);
 }
